--- a/python/sglang/srt/layers/linear.py
+++ b/python/sglang/srt/layers/linear.py
@@ -1174,7 +1174,7 @@ class RowParallelLinear(LinearBase):
 
         self.input_is_parallel = input_is_parallel
         self.reduce_results = reduce_results
-
+        self.chunk_size = 5120
         # Divide the weight matrix along the last dimension.
         if tp_rank is None:
             tp_rank = get_tensor_model_parallel_rank()
@@ -1275,27 +1275,61 @@ class RowParallelLinear(LinearBase):
             param.load_row_parallel_weight(loaded_weight)
 
     def forward(self, input_):
-        if self.input_is_parallel:
-            input_parallel = input_
-        else:
-            splitted_input = split_tensor_along_last_dim(
-                input_, num_partitions=self.tp_size
-            )
-            input_parallel = splitted_input[self.tp_rank].contiguous()
+        if input_.size(0) >= 16384:
+            self.chunk_size = input_.size(0)//3
+        num_chunks = (input_.size(0) + self.chunk_size - 1)//self.chunk_size
+        if num_chunks <2:
+            if self.input_is_parallel:
+                input_parallel = input_
+            else:
+                splitted_input = split_tensor_along_last_dim(
+                    input_, num_partitions=self.tp_size
+                )    
+                input_parallel = splitted_input[self.tp_rank].contiguous()
+            assert self.quant_method is not None
+            # Only fuse bias add into GEMM for rank 0 (this ensures that
+            # bias will not get added more than once in TP>1 case)
+            bias_ = None if (self.tp_rank > 0 or self.skip_bias_add) else self.bias
+            output_parallel = self.quant_method.apply(self, input_parallel, bias=bias_)
+            if self.reduce_results and self.tp_size > 1:
+                output = tensor_model_parallel_all_reduce(output_parallel)
+            else:
+                output = output_parallel
 
-        # Matrix multiply.
-        assert self.quant_method is not None
-        # Only fuse bias add into GEMM for rank 0 (this ensures that
-        # bias will not get added more than once in TP>1 case)
-        bias_ = None if (self.tp_rank > 0 or self.skip_bias_add) else self.bias
-        output_parallel = self.quant_method.apply(self, input_parallel, bias=bias_)
-        if self.reduce_results and self.tp_size > 1:
-            output = tensor_model_parallel_all_reduce(output_parallel)
-        else:
-            output = output_parallel
+            output_bias = self.bias if self.skip_bias_add else None
+            return output, output_bias
+            
+        output_shape = [input_.size(0),self.output_size]
+        output = torch.empty(*output_shape,dtype = input_.dtype,device = input_.device)
+        comm_stream = torch.cuda.Stream()
+        compute_events = [torch.cuda.Event(enable_timing=False) for _ in range(num_chunks)]
+        comm_events = [torch.cuda.Event(enable_timing=False) for _ in range(num_chunks)]
+        for i in range(num_chunks):
+            chunk_start = i*self.chunk_size
+            chunk_end = min((i+1)*self.chunk_size,input_.size(0))
+            input_chunk = input_[chunk_start:chunk_end]
+            if self.input_is_parallel:
+                input_parallel = input_chunk
+            else:
+                splitted_input = split_tensor_along_last_dim(input_chunk,num_partitions=self.tp_size)
+                input_parallel = splitted_input[self.tp_rank].contiguous()
 
-        output_bias = self.bias if self.skip_bias_add else None
-
+            assert self.quant_method is not None
+            bias_ = None if (self.tp_rank > 0 or self.skip_bias_add) else self.bias
+            output_parallel = self.quant_method.apply(self, input_parallel, bias=bias_)
+            compute_events[i].record()
+            if self.reduce_results and self.tp_size > 1:
+                with torch.cuda.stream(comm_stream):
+                    comm_stream.wait_event(compute_events[i])
+                    output_chunk = tensor_model_parallel_all_reduce(output_parallel)
+                    output[chunk_start:chunk_end,:] = output_chunk
+                    comm_events[i].record()
+            else:
+                output[chunk_start:chunk_end] = output_parallel
+                if i<num_chunks-1:
+                    torch.cuda.current_stream().wait_event(compute_events[i])
+        if self.reduce_results and self.tp_size >1:            
+            torch.cuda.current_stream().wait_event(comm_events[-1])        
         return output, output_bias
 
     def extra_repr(self) -> str: