--- a/python/sglang/srt/layers/linear.py
+++ b/python/sglang/srt/layers/linear.py
@@ -1329,7 +1329,8 @@ class RowParallelLinear(LinearBase):
                 if i<num_chunks-1:
                     torch.cuda.current_stream().wait_event(compute_events[i])
         if self.reduce_results and self.tp_size >1:            
-            torch.cuda.current_stream().wait_event(comm_events[-1])        
+            torch.cuda.current_stream().wait_event(comm_events[-1])
+        output_bias = self.bias if self.skip_bias_add else None
         return output, output_bias
 
     def extra_repr(self) -> str: