Coarse-grained computations-communicate hidden

## Motivation

Compute-communication hiding by means of chunking and cuda stream to improve throughput.

## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [x] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [x] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.
## Experiment setup
### 4×H200
### command:
```
export CUDA_VISIBLE_DEVICES=4,5,6,7
model_path=/models/Qwen2.5-72B-Instruct-FP8-PER-TENSOR
dataset_path=/data/ShareGPT_V3_unfiltered_cleaned_split.json
python -m sglang.bench_offline_throughput \
        --model-path ${model_path} \
        --dataset-name random \
        --random-input 1024 \
        --random-output 1 \
        --num-prompts 1000 \
        --dataset-path  ${dataset_path} \
        --chunked-prefill-size 16384 \
        --tp-size 4
```

## Results
### The baseline is SGLang Release v0.4.6
 Repeated 5 times,  get the following results for the throughput test(by the bench_offline_throughput.py)：
Baseline: 21457.22 tok/s, 21460.65 tok/s, 21452.04 tok/s, 21460.67 tok/s, 21392.38 tok/s
pr: 21854.43 tok/s, 21846.43 tok/s, 21789.11 tok/s, 21831.50 tok/s, 21784.51 tok/s
#### average:
baseline:  21444.592 tok/s
pr:   21821.196 tok/s